requirements.txt
blinker==1.6.2
click==8.1.7
Flask==3.0.0
Flask-Testing==0.8.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-decouple==3.8
python-dotenv==1.0.0
Werkzeug==3.0.0
